refactor(api): build httpClient.ts + split learningApi into domain modules (Related to #1827)
Phase 1 — httpClient.ts (80 lines):
- Centralized fetch wrapper: auto-auth header, JSON parse, 401 → SessionExpiredApiError,
  204 → undefined, typed ApiError on failures
- Helper shortcuts: get/post/put/patch/del

Phase 2 — learning/ domain split (1031-line grab-bag → 7 focused modules):
- session.ts (411): createLearningSession, evaluateReading, session status/list/report,
  student progress, step persistence, reading history
- comprehension.ts (256): socratic chat, AI reading analysis, scoring, dialogue history
- listening.ts (40): listening comprehension evaluation
- dictionary.ts (59): dictionary lookup
- exitTicket.ts (86): exit ticket submit/retrieve
- sentence.ts (48): sentence practice
- mcq.ts (125): MCQ rescue AI tutor + attempt telemetry
- index.ts: re-exports all + SessionExpiredError from api.ts

learningApi.ts becomes 3-line compatibility shim (all callers unchanged).
Vite build passes in 6.87s. Zero new TypeScript errors.
