+++ frontend/src/services/learning/exitTicket.ts
@@ -1,0 +1,86 @@
+import { API_BASE } from '../apiConfig';
+
+// ---------------------------------------------------------------------------
+// Exit ticket (Issue #463)
+// ---------------------------------------------------------------------------
+
+export interface ExitTicketQuestion {
+  id: number;
+  question: string;
+  options: string[];
+  correct_index: number;
+  explanation: string;
+}
+
+export interface ExitTicketGenerateResponse {
+  questions: ExitTicketQuestion[];
+  /**
+   * 'pregenerated' = served from lesson YAML multiple_choice (no AI call, Issue #1402)
+   * 'ai' = generated by Gemini at request time (legacy fallback)
+   * 'fallback' = empty list, frontend should use local rule-based questions
+   */
+  source: 'pregenerated' | 'ai' | 'fallback';
+}
+
+export interface ExitTicketSubmitResponse {
+  score: number;
+  correct_count: number;
+  total: number;
+  passed: boolean;
+}
+
+export async function generateExitTicket(
+  token: string,
+  sessionId: number,
+  storyContent: string[],
+  wrongChars?: string[],
+): Promise<ExitTicketGenerateResponse> {
+  const res = await fetch(
+    `${API_BASE}/api/learning/sessions/${sessionId}/exit-ticket/generate`,
+    {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        story_content: storyContent,
+        wrong_chars: wrongChars ?? [],
+      }),
+    },
+  );
+  if (!res.ok) {
+    const body = await res.json().catch(() => ({ detail: 'Unknown error' }));
+    throw new Error(body.detail || `generateExitTicket failed: ${res.status}`);
+  }
+  return res.json();
+}
+
+export async function submitExitTicket(
+  token: string,
+  sessionId: number,
+  answers: Array<{ question_id: number; selected_index: number }>,
+  score: number,
+  totalQuestions: number,
+): Promise<ExitTicketSubmitResponse> {
+  const res = await fetch(
+    `${API_BASE}/api/learning/sessions/${sessionId}/exit-ticket/submit`,
+    {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${token}`,
+        'Content-Type': 'application/json',
+      },
+      body: JSON.stringify({
+        answers,
+        score,
+        total_questions: totalQuestions,
+      }),
+    },
+  );
+  if (!res.ok) {
+    const body = await res.json().catch(() => ({ detail: 'Unknown error' }));
+    throw new Error(body.detail || `submitExitTicket failed: ${res.status}`);
+  }
+  return res.json();
+}
